pyproject.toml
[project]
name = "mneme"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "2mal3", email = "56305732+2mal3@users.noreply.github.com" }
]
dependencies = [
    "discord-py==2.*",
    "python-dotenv==1.*",
    "pydantic==2.*",
    "niquests==3.*",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.ruff]
target-version = "py311"
line-length = 120

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["mneme"]
